LayoutTests/inspector/layers/layers-compositing-reasons-expected.txt


=== Enable the LayerTree agent ===

PASS

=== Get the Document ===

PASS

=== Get the layer tree ===

PASS

=== Check layers' reasons for compositing ===

PASS: <div id="opacity-container"> is composited due to having an opacity style and a composited child.
PASS: <div id="child"> is composited due to having "backface-visibility: hidden" and a 3D transform.
PASS: <canvas id="canvas"> is composited due to having a 3D transform.
